+++ src/java/com/enseval/ttss/vm/PopBuatInvoiceVM.java
@@ -46,7 +46,7 @@ public class PopBuatInvoiceVM {
     private Window winBuatInvoice;
     User userLogin;
     List<InvoiceItem> listInvoiceItem = new ArrayList<>();
-    Long totalHarga = 0L;
+    double totalHarga = 0;
 
     Invoice invoice;
 
@@ -215,7 +215,7 @@ public class PopBuatInvoiceVM {
     @Command
     @NotifyChange({"totalHarga"})
     public void doCount() {
-        this.totalHarga = 0L;
+        this.totalHarga = 0;
         for (InvoiceItem item : listInvoiceItem) {
             this.totalHarga += item.getHargaSatuan() * item.getJmlKemasan();
         }
@@ -246,13 +246,15 @@ public class PopBuatInvoiceVM {
         this.listInvoiceItem = listInvoiceItem;
     }
 
-    public Long getTotalHarga() {
+    public double getTotalHarga() {
         return totalHarga;
     }
 
-    public void setTotalHarga(Long totalHarga) {
+    public void setTotalHarga(double totalHarga) {
         this.totalHarga = totalHarga;
     }
+
+    
 
     public Invoice getInvoice() {
         return invoice;
